+++ Mods ChoGGi/Specialist by Experience/metadata.lua
@@ -12,7 +12,7 @@ return PlaceObj("ModDef", {
 	"lua_revision", LuaRevision,
   "description", [[Colonists that work at the same job for 25 Sols will automagically get the specialisation needed for that workplace.
 
-Does not include colonists that already have a spec (you can allow that with ModConfig).
+Does not include colonists that already have a spec (you can allow that with Mod Config Reborn).
 Works on saved games, but only on newly hired colonists (you can use ECM to fire everyone).
 
 Idea by Dragonmystic.]],