+++ ChoGGi's Library/Changelog.md
@@ -4,6 +4,7 @@
 
 ## v4.6 (06 Dec 2018)
 - Midly broke list choice filter (for certain last items).
+- Updates dialog scale when user changes scale in Options.
 
 ## v4.5 (06 Dec 2018)
 - Removed length limit of dialog titles.